fix(cli): remove mcp index/tool import cycle
