+++ packages/cli/src/mcp/formatters.ts
@@ -1,4 +1,9 @@
 import type { Memory, MemoryType } from "../lib/memory.js";
+import {
+  formatStorageWarningsForText,
+  getStorageWarnings,
+  type StorageWarning,
+} from "../lib/storage-health.js";
 
 // ─── Type Labels ──────────────────────────────────────────────────────────────
 
@@ -42,4 +47,37 @@ export interface ToolResponsePayload {
   content: ToolTextPart[];
   isError?: boolean;
   [key: string]: unknown;
+}
+
+// ─── Storage Warning Helper ───────────────────────────────────────────────────
+
+export async function withStorageWarnings(
+  result: ToolResponsePayload,
+  warningsOverride?: StorageWarning[]
+): Promise<ToolResponsePayload> {
+  if (result.isError) return result;
+
+  if (result.content.length === 0) return result;
+
+  try {
+    const warnings = warningsOverride ?? (await getStorageWarnings()).warnings;
+    if (warnings.length === 0) return result;
+
+    const warningBlock = formatStorageWarningsForText(warnings);
+    if (!warningBlock) return result;
+
+    const nextContent = [...result.content];
+    const textPart = nextContent[0];
+    nextContent[0] = {
+      ...textPart,
+      text: `${textPart.text}\n\n${warningBlock}`,
+    };
+
+    return {
+      ...result,
+      content: nextContent,
+    };
+  } catch {
+    return result;
+  }
 }
